+++ changelog.md
@@ -1,8 +1,15 @@
-# 0.1.11
+# 0.2.0
+- Added
+  - Retain last N versions of a mod.
+    - Keep only the latest version of a mod by setting this to 1.
+  - Option to keep the same folder name when updating a mod.
+    - This is the new default. It is useful for modders who create a dependency on a folder path and don't want that path to change. 
 - Fixed
   - Self-update failed if there were any non-ASCII characters in the TriOS or temp folder paths.
     - The Windows self-updater now works completely differently, doing an in-place update with no .bat script.
     - Windows, you so crazy.
+- Changed
+  - Incompatible mods are now dimmed on the Dashboard.
 
 # 0.1.10
 - Fixed